+++ CHANGELOG.md
@@ -3,7 +3,12 @@
 
 ## [Unreleased]
 ### Added
+- add `ixf_last_import` to InternetExchangeBase (peeringdb/peeringdb#683)
+- add `ixf_net_count` to InternetExchangeBase (peeringdb/peeringdb#683)
+- add `ixf_ixp_member_list_url` to IXLanBase (peeringdb/peeringdb#249)
+- add `ixf_ixp_member_list_url_visible` to IXLanBase (peeringdb/peeringdb#249)
 ### Fixed
+- fix issue in setup.py for test requirements
 ### Changed
 ### Deprecated
 ### Removed